created payment controllers + routes fixed

+++ app/models/Payment.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Payment(id: Int = 0, order_id: Int, date: String, amount: Float)
+case class Payment(id: Int = 0, order_id: Int, date: String, amount: Int)
 
 object Payment {
   implicit val paymentFormat: OFormat[Payment] = Json.format[Payment]
